myjourney/requirements/local.txt
-r base.txt

black==21.5b2
flake8==3.9.2
pytest-django==4.4.0